sensor3Dcartesian added to sensor base class and init calls tided up for child sensors
