+++ stonesoup/sensor/passive.py
@@ -2,15 +2,14 @@
 import copy
 import numpy as np
 
-from .base import Sensor
+from .base import Sensor3DCartesian
 from ..base import Property
 from ..models.measurement.nonlinear import CartesianToElevationBearing
 from ..types.array import CovarianceMatrix
 from ..types.detection import Detection
-from ..types.state import State, StateVector
 
 
-class PassiveElevationBearing(Sensor):
+class PassiveElevationBearing(Sensor3DCartesian):
     """A simple passive sensor that generates measurements of targets, using a
     :class:`~.CartesianToElevationBearing` model, relative to its position.
 
@@ -20,18 +19,8 @@ class PassiveElevationBearing(Sensor):
 
     """
 
-    position = Property(StateVector,
-                        doc="The sensor position on a 3D Cartesian plane,\
-                             expressed as a 3x1 array of Cartesian coordinates\
-                             in the order :math:`x,y,z`")
-    orientation = Property(
-        StateVector,
-        doc="A 3x1 array of angles (rad), specifying the sensor orientation in \
-            terms of the counter-clockwise rotation around each Cartesian \
-            axis in the order :math:`x,y,z`. The rotation angles are positive \
-            if the rotation is in the counter-clockwise direction when viewed \
-            by an observer looking along the respective rotation axis, \
-            towards the origin")
+    measurement_model = None
+
     ndim_state = Property(
         int,
         doc="Number of state dimensions. This is utilised by (and follows in\
@@ -47,25 +36,16 @@ class PassiveElevationBearing(Sensor):
                                 :class:`~.CartesianToElevationBearing`\
                                 model")
 
-    def __init__(self, position, orientation, ndim_state, mapping, noise_covar,
-                 *args, **kwargs):
-        measurement_model = CartesianToElevationBearing(
-            ndim_state=ndim_state,
-            mapping=mapping,
-            noise_covar=noise_covar,
-            translation_offset=position,
-            rotation_offset=orientation)
+    def __init__(self, *args, **kwargs):
 
-        super().__init__(position, orientation, ndim_state, mapping,
-                         noise_covar, *args, measurement_model, **kwargs)
+        super().__init__(*args, **kwargs)
 
-    def set_position(self, position):
-        self.position = position
-        self.measurement_model.translation_offset = position
-
-    def set_orientation(self, orientation):
-        self.orientation = orientation
-        self.measurement_model.rotation_offset = orientation
+        self.measurement_model = CartesianToElevationBearing(
+            ndim_state=self.ndim_state,
+            mapping=self.mapping,
+            noise_covar=self.noise_covar,
+            translation_offset=self.position,
+            rotation_offset=self.orientation)
 
     def gen_measurement(self, ground_truth, noise=None, **kwargs):
         """Generate a measurement for a given state
